# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FleetGuard API — Inspecciones DVIR (CAM-11), login (CAM-43), mantenimiento preventivo (CAM-40) programación (CAM-42) y órdenes de trabajo (CAM-14, CAM-60)
  description: >
    Contrato para la carga de inspecciones pre-trip / post-trip del chofer y
    para la consulta de los defectos que generan (CAM-13). Cubre el listado
    de vehículos, la subida de fotos de defectos, el envío de la inspección
    y el listado de defectos, además del login (CAM-43). No cubre roles /
    autorización por rol todavía, ni las órdenes de trabajo de mantenimiento
    (son consumidoras de este listado, no parte de él).


    También cubre el mantenimiento preventivo y el componente "Estado de la
    flota" del dashboard de admin (CAM-40): catálogo de planes de
    mantenimiento, asignación de planes a vehículos, registro de
    completions, y la vista agregada de GET /api/vehicles?view=fleet-status.
    Ver docs/api/CAM-40-fleet-status-contract.md para el detalle de
    decisiones de diseño.


    Y la programación de mantenimientos y arreglos de defectos (CAM-42,
    CAM-50, CAM-51): cuándo se planea hacer un mantenimiento o resolver un
    defecto, distinto de cuándo corresponde (calculado) y de cuándo ya se
    hizo (historial). Ver docs/api/CAM-42-schedule-contract.md.


    Y las órdenes de trabajo (CAM-14, CAM-15, CAM-62, CAM-63), con el
    técnico a cargo (CAM-60) y el listado de usuarios de solo lectura que
...
  version: "1.6"
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/vehicles/{vehicleId}/checklist:
    get:
      summary: Checklist DVIR efectivo de un vehículo (CAM-31)
      description: >
        Lo que el chofer tiene que completar para este vehículo. pre-trip =
        catálogo base − ítems desactivados + ítems extra del vehículo; post-trip
        es siempre la lista fija. Ver CAM-31-vehicle-checklist-contract.md.
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [pre-trip, post-trip]
            default: pre-trip
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChecklistItem"
        "404":
          description: VEHICLE_NOT_FOUND
        "422":
          $ref: "#/components/responses/FieldValidationErrorResponse"
  /api/vehicles/{vehicleId}/checklist-items:
    get:
      summary: Configuración del checklist pre-trip de un vehículo (admin, CAM-31)
      description: >
        Todos los ítems del catálogo base (con enabled según el vehículo) más los
        ítems extra vigentes. locked=true son obligatorios y no se pueden quitar.
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VehicleChecklistItem"
        "404":
          description: VEHICLE_NOT_FOUND
    post:
      summary: Agregar un ítem extra al checklist pre-trip del vehículo (CAM-31)
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleChecklistItemRequest"
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleChecklistItem"
        "404":
          description: VEHICLE_NOT_FOUND
        "422":
          $ref: "#/components/responses/FieldValidationErrorResponse"
  /api/vehicles/{vehicleId}/checklist-items/{itemId}:
    patch:
      summary: Quitar o volver a activar un ítem del checklist del vehículo (CAM-31)
      description: >
        Sirve para ítems base (se guardan como desactivados para este vehículo) y
        para ítems extra (baja lógica). Las inspecciones viejas no cambian.
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleChecklistItemRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleChecklistItem"
        "404":
          description: VEHICLE_NOT_FOUND / CHECKLIST_ITEM_NOT_FOUND
        "409":
          description: ITEM_LOCKED -- el ítem es obligatorio (kilómetros actuales)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          $ref: "#/components/responses/FieldValidationErrorResponse"
  /api/vehicles/{id}:
    get:
      summary: Detalle de un vehículo (CAM-22)
      description: Mismo shape que un item de `GET /api/vehicles`. Devuelve también vehículos dados de baja.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleSummary"
        "404":
          description: VEHICLE_NOT_FOUND
    patch:
      summary: Editar un vehículo, o reactivarlo (CAM-25)
      description: >
        Campos ausentes no se modifican. `active: true` reactiva un vehículo
        dado de baja. Si el vehículo está dado de baja, el único PATCH
        permitido es `{ "active": true }` solo -- cualquier otro campo
        (junto con la reactivación o sin ella) devuelve `409 VEHICLE_INACTIVE`.
        Ver CAM-25-vehicle-abm-contract.md.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleSummary"
        "404":
          description: VEHICLE_NOT_FOUND
        "409":
          description: >
            DUPLICATE_PLATE -- si se cambia la patente a una que ya existe.
            VEHICLE_INACTIVE -- si se edita cualquier campo de un vehículo
            dado de baja sin reactivarlo primero.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                duplicatePlate:
                  value: { error: "DUPLICATE_PLATE", message: "Ya existe un vehículo con la patente AB123CD" }
                vehicleInactive:
                  value: { error: "VEHICLE_INACTIVE", message: "El vehículo está dado de baja -- reactivalo primero con PATCH { \"active\": true }" }
        "422":
          $ref: "#/components/responses/FieldValidationErrorResponse"
    delete:
      summary: Dar de baja un vehículo (CAM-25)
      description: >
        Baja lógica: pone `active=false`, no borra la fila ni su historial
        (inspecciones, defectos, asignaciones de mantenimiento). Ver
        CAM-25-vehicle-abm-contract.md.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Dado de baja
        "404":
          description: VEHICLE_NOT_FOUND
        "409":
          description: VEHICLE_ON_TRIP -- el vehículo tiene un viaje abierto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    FieldValidationErrorResponse:
      description: Datos inválidos (CAM-40) -- details[] usa "field", no "itemId".
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FieldValidationError"
  schemas:
    VehicleStatus:
      type: string
      enum: [available, on-trip]
    VehicleSummary:
      type: object
      required: [id, plate, brand, model, status, odometerKm, active]
      properties:
        id: { type: string }
        plate: { type: string, example: "AB123CD" }
        brand: { type: string, example: "Mercedes-Benz" }
        model: { type: string, example: "Sprinter" }
        status:
          $ref: "#/components/schemas/VehicleStatus"
        vehicleType: { type: string, nullable: true, example: "furgon" }
        year: { type: integer, nullable: true, example: 2022 }
        chassisNumber: { type: string, nullable: true }
        odometerKm: { type: integer }
        active:
          type: boolean
          description: >
            false si el vehículo está dado de baja (CAM-25) -- ver el
            parámetro `active` de GET /api/vehicles.
...
    UpdateVehicleRequest:
      type: object
      description: Campos ausentes (null) no se modifican.
      properties:
        plate: { type: string }
        brand: { type: string }
        model: { type: string }
        vehicleType: { type: string }
        year: { type: integer }
        chassisNumber: { type: string }
        active:
          type: boolean
          description: En false da de baja; en true reactiva (ver CAM-25-vehicle-abm-contract.md).
...
    ChecklistItem:
      type: object
      required: [id, label, type, section, required]
      properties:
        id: { type: string, description: "Id del catálogo base (ej. ext-luces) o extra-<uuid> si es un ítem extra del vehículo." }
        label: { type: string }
        type: { type: string, enum: [check, number] }
        section: { type: string, enum: [exterior, interior, posttrip] }
        required: { type: boolean }
    VehicleChecklistItem:
      allOf:
        - $ref: "#/components/schemas/ChecklistItem"
        - type: object
          required: [origin, enabled, locked]
          properties:
            origin: { type: string, enum: [base, extra] }
            enabled: { type: boolean }
            locked: { type: boolean, description: Obligatorio, no se puede quitar. }
    CreateVehicleChecklistItemRequest:
      type: object
      required: [label, type, section]
      properties:
        label: { type: string, minLength: 1, maxLength: 60 }
        type: { type: string, enum: [check, number] }
        section: { type: string, enum: [exterior, interior] }
    UpdateVehicleChecklistItemRequest:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
...
    FieldValidationError:
      allOf:
        - $ref: "#/components/schemas/ApiError"
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }

    # --- Mantenimiento preventivo y estado de flota (CAM-40) ---
...